Move script sources to slug-named directories
