+++ src/userscripts/freee-shortcuts/index.user.ts
