+++ src/userscripts/github-auto-sso/index.user.ts
