+++ src/userscripts/github-disable-some-keymaps/index.user.ts
@@ -1,4 +1,4 @@
-import { isTyping } from "./utils/isTyping";
+import { isTyping } from "../utils/isTyping";
 
 document.body.addEventListener("keydown", (e) => {
   if (isTyping()) return;
