+++ src/userscripts/github-pr-submission-shortcuts/index.user.ts
@@ -1,4 +1,4 @@
-import { isTyping } from "./utils/isTyping";
+import { isTyping } from "../utils/isTyping";
 
 void (() => {
   window.addEventListener("keydown", (e) => {
