+++ src/userscripts/google-calendar-unintended-user-alert/index.user.ts
