+++ src/userscripts/mdn-force-english/index.user.ts
