+++ src/userscripts/moneytree-shortcuts/index.user.ts
