+++ src/userscripts/pinterest-save-from-site-shortcuts/index.user.ts
