+++ src/userscripts/scrapbox-clear-watch-list/index.user.ts
