+++ src/userscripts/scrapbox-force-theme/index.user.ts
