+++ src/userscripts/scrapbox-no-project-styles/index.user.ts
