+++ src/userscripts/scrapbox-project-notice-in-new-page/index.user.ts
