+++ src/userscripts/slack-no-autofocus/index.user.ts
