+++ src/userscripts/slack-no-unintended-reload/index.user.ts
